Whole matrix of data types in code and tests

+++ java/cuvs-java/src/main/java22/com/nvidia/cuvs/internal/HostMemoryDatasetImpl.java
@@ -156,7 +156,7 @@ public class HostMemoryDatasetImpl extends DatasetBaseImpl {
 
     @Override
     public float getAsFloat(long index) {
-      assert dataType == DataType.INT;
+      assert dataType == DataType.FLOAT;
       return memorySegment.get((ValueLayout.OfFloat) valueLayout, index * valueByteSize);
     }
